requirements.txt
mistralai==2.1.3
python-dotenv>=1.0.0